Add desktop navbar.
